Check datatype and force in replaybuffer

+++ baselines/deepq/replay_buffer.py
@@ -35,10 +35,10 @@ class ReplayBuffer(object):
         for i in idxes:
             data = self._storage[i]
             obs_t, action, reward, obs_tp1, done = data
-            obses_t.append(np.array(obs_t, copy=False))
+            obses_t.append(np.array(obs_t, copy=False, dtype=np.uint8))
             actions.append(np.array(action, copy=False))
             rewards.append(reward)
-            obses_tp1.append(np.array(obs_tp1, copy=False))
+            obses_tp1.append(np.array(obs_tp1, copy=False, dtype=np.uint8))
             dones.append(done)
         return np.array(obses_t), np.array(actions), np.array(rewards), np.array(obses_tp1), np.array(dones)
